Condense list items into parent list component
