+++ client/src/components/ManageCards/ManageCards.tsx
@@ -2,9 +2,9 @@ import React, { useState, useEffect } from 'react';
 import { Card } from '../../../../types';
 import { ColouredButton, ButtonType } from '../ColouredButton/ColouredButton';
 import { CardForm, CardFormType } from '../CardForm/CardForm';
-import { getAllCards } from '../../utils/axios';
-import CardListItem from '../ManageCardsListItem/ManageCardsListItem';
+import { getAllCards, deleteCard } from '../../utils/axios';
 import Divider from '../Divider/Divider';
+
 
 const ManageCards = () => {
   const [addCardVisisble, setAddCardVisisble] = useState(false);
@@ -55,7 +55,7 @@ const ManageCards = () => {
       return cards.map((card: Card) => {
         return ( 
           <div key={card._id}>
-            <CardListItem 
+            <ManageCardsListItem 
               id={card._id} 
               prompt={card.prompt}
               answer={card.answer}
@@ -87,5 +87,62 @@ const ManageCards = () => {
     </div>
   );
 }
+
+type ManageCardsListItemProps = {
+  id: string
+  prompt: string
+  answer: string
+  onEdit: () => void
+  onDelete: () => void
+}
+
+const ManageCardsListItem = (props: ManageCardsListItemProps) => {
+  const [editMode, setEditMode] = useState(false);
+  const id = props.id;
+  const prompt = props.prompt;
+  const answer = props.answer;
+  
+  const renderItemContent = () => {
+    if(editMode){
+      return (
+        <CardForm
+          afterGreenButtonClick={async () => {
+            await props.onEdit() 
+            setEditMode(false);
+          }}
+          onCancel={() => setEditMode(false)}
+          formType={CardFormType.edit}
+          cardId={id}
+          cardPrompt={prompt}
+          cardAnswer={answer}
+        />
+      );
+    }
+    else{
+      return(
+        <div>
+          <p>Prompt: {prompt} </p>
+          <p>Answer: {answer}</p>
+        </div>
+      );
+    }
+  }
+
+  return (
+    <div data-testid={`manage-card-list-item-${props.id}`}>
+      <ColouredButton onClickAction={() => { setEditMode(true) }} text="edit" buttonType={ButtonType.default} />
+      <ColouredButton 
+        onClickAction={async () => { 
+          await deleteCard(id);
+          props.onDelete();
+        }} 
+        text="delete" 
+        buttonType={ButtonType.delete} 
+      />
+      {renderItemContent()}
+    </div>
+  );
+}
+
 
 export default ManageCards;
